Rename repo-task-store to repo-tasks-store
